requirements.txt
flask
python-dotenv
openai
flask_sqlalchemy
conda